requirements.txt
openai==0.28.1
pydantic==1.10.12
python-dotenv==1.0.0
termcolor==2.4.0
tiktoken==0.6.0